backend/requirements-local.txt
-r requirements.txt
faster-whisper
pyannote.audio
torch
langdetect==1.0.9
